+++ back-end/test-rules/jest.config.cjs
@@ -5,6 +5,9 @@ module.exports = {
   testEnvironment: 'jest-environment-node',
   transform: {},
 
+  // Needed for using 'firebase-jest-testing', until Jest resolver supports modules with 'exports' (see TRACK.md).
+  resolver: "../hack-jest/custom-resolver.cjs",
+
   testRunner: "jest-circus/runner",   // upcoming default for Jest (should be faster/better...)
 
   // Default is 5000. None of our tests take that long; fail fast.